+++ tests/test_class_schema.py
@@ -1,9 +1,9 @@
+import typing
 import unittest
 from typing import Any
 from uuid import UUID
 
 import dataclasses
-import typing
 from marshmallow import Schema, ValidationError
 from marshmallow.fields import Field, UUID as UUIDField, List as ListField, Integer
 
@@ -67,6 +67,51 @@ class TestClassSchema(unittest.TestCase):
         self.assertIsInstance(schema.fields["custom_list"], CustomListField)
         self.assertIsInstance(schema.fields["uuid"], UUIDField)
         self.assertIsInstance(schema.fields["n"], Integer)
+
+    def test_filtering_list_schema(self):
+        class FilteringListField(ListField):
+            def __init__(
+                self,
+                cls_or_instance: typing.Union[Field, type],
+                min: typing.Any,
+                max: typing.Any,
+                **kwargs,
+            ):
+                super().__init__(cls_or_instance, **kwargs)
+                self.min = min
+                self.max = max
+
+            def _deserialize(
+                self, value, attr, data, **kwargs
+            ) -> typing.List[typing.Any]:
+                loaded = super()._deserialize(value, attr, data, **kwargs)
+                return [x for x in loaded if self.min <= x <= self.max]
+
+        class BaseSchema(Schema):
+            TYPE_MAPPING = {typing.List: FilteringListField}
+
+        @dataclasses.dataclass
+        class WithCustomField:
+            constrained_floats: typing.List[float] = dataclasses.field(
+                metadata={"max": 2.5, "min": 1}
+            )
+            constrained_strings: typing.List[str] = dataclasses.field(
+                metadata={"max": "x", "min": "a"}
+            )
+
+        schema = class_schema(WithCustomField, base_schema=BaseSchema)()
+        actual = schema.load(
+            {
+                "constrained_floats": [0, 1, 2, 3],
+                "constrained_strings": ["z", "a", "b", "c", ""],
+            }
+        )
+        self.assertEqual(
+            actual,
+            WithCustomField(
+                constrained_floats=[1.0, 2.0], constrained_strings=["a", "b", "c"]
+            ),
+        )
 
     def test_any_none(self):
         # See: https://github.com/lovasoa/marshmallow_dataclass/issues/80